+++ build.bat
@@ -1,2 +1,2 @@
-pyinstaller Commander_Tracker.py --onefile -w --collect-all="tksvg"
+pyinstaller main.py --onefile -w --collect-all="tksvg"
 pause
